refactor: create pathutil package
